Add files via upload
in this project, i took existing data about weather and how that affects the usage of rental bikes in the area. i split the entire data set into train and test data, and at the end tested the model on the test data. then i compared the results of the test to already existing data to see how accurate my model was.